// src/modules/user/enums/messages/user/user.messages.enum.ts
export enum UserMessage {
  NOT_FOUND = 'User not found',
  DATA_REQUIRED = 'User data required',
  REQUIRED = 'User required',
  ID_REQUIRED = 'User id is required',

  // user id
  REQUIRED_USER_ID = 'User id is required',
  USER_ID_TYPE = 'User id should be integer or equal 1',
  INVALID_USER_ID = 'Invalid user id',
  NULL_USER_ID = 'Null user id',
}